[WebDriver][win] make the `New Window` command duplicates an application window, instead of creating new one. https://bugs.webkit.org/show_bug.cgi?id=289887
Reviewed by NOBODY (OOPS!).

Before this change, the WebDriver creates new WebPageProxy and WebView for the `New Window` command.
This method is simple but not appropriate for testing the application, as the newly created window
differs from the application's.
The purpose of this change is to solve the above problem.

The main change in this commit is that the AutomationSessionClient will no longer create a WebView
when it receives `New Window` command, but will instead call the `WKPageUIClient::createNewPage`.
The creation of the window is delegated to the application, which will allow for more correct testing.

* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/UIProcess/win/AutomationClientWin.cpp:
(WebKit::AutomationSessionClient::requestNewPageWithOptions):
(WebKit::AutomationClient::closeAutomationSession):
(WebKit::AutomationSessionClient::close): Deleted.
(WebKit::AutomationSessionClient::didReceiveAuthenticationChallenge): Deleted.
(WebKit::AutomationSessionClient::retainWebView): Deleted.
(WebKit::AutomationSessionClient::releaseWebView): Deleted.
* Source/WebKit/UIProcess/win/AutomationClientWin.h:
* Source/WebKit/UIProcess/win/WebPageProxyWin.cpp:
(WebKit::WebPageProxy::requestNewWindow):
* Tools/MiniBrowser/win/WebKitBrowserWindow.cpp:
(WebKitBrowserWindow::WebKitBrowserWindow):
(WebKitBrowserWindow::didReceiveAuthenticationChallenge):
(WebKitBrowserWindow::close):
* Tools/MiniBrowser/win/WebKitBrowserWindow.h: