*** This commit is only for internal review. ***
Changes:

https://github.com/WebKit/WebKit/pull/42605/files#r2006479179
I thought that the "create window" API was inappropriate for WebPageProxy, so I changed the window creation to AutomationSessionClient.

https://github.com/WebKit/WebKit/pull/42605/files#r2006476663
In win port, webdriver uses the same protocol as inspector, and it is okay to assume that there is always a webpage to which the socket is connected.

https://github.com/WebKit/WebKit/pull/42605/files#r2006212370
Added the comment.

https://github.com/WebKit/WebKit/pull/42605/files#r2006209549
Added the FIXME comment. I'll separate .cpp and .h files after landing this pullrequest.

https://github.com/WebKit/WebKit/pull/42605/files#r2006201565
Added the bugzilla link.

https://github.com/WebKit/WebKit/pull/42605/files#r2006453963
I think it is not necessary to distinguish where this “createNewPage” request is called from, or to customize the client behavior.